Moved add item list to List view, added paragraph explaining purchase urgency, added name of list in h2, commented out add list form code in Manage list view
